+++ apps/dash-lyft-explorer/assets/css/header.css
@@ -1,0 +1,51 @@
+.header {
+    padding: 5px 15px 5px 10px;
+}
+
+.header-logos {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: flex-end;
+    gap: 5%;
+}
+
+.app-body {
+    padding: 20px 10px 10px 10px;
+}
+
+
+/* Demo button css */
+.demo-button {
+    font-family: Open Sans,sans-serif;
+    text-decoration: none;
+    display: inline-flex;
+    -webkit-align-items: center;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    border-radius: 8px;
+    font-weight: 700;
+    height: 2.5rem;
+    font-size: 13px;
+    -webkit-padding-start: 1rem;
+    padding-inline-start: 1rem;
+    -webkit-padding-end: 1rem;
+    padding-inline-end: 1rem;
+    color: #ffffff;
+    letter-spacing: 2px;
+    border: solid 1.5px transparent;
+    box-shadow: 2px 1000px 1px #0c0c0c inset;
+    background-image: linear-gradient(135deg, #7A76FF, #7A76FF, #7FE4FF);
+    -webkit-background-size: 200% 100%;
+    background-size: 200% 100%;
+    -webkit-background-position: 99%;
+    background-position: 99%;
+    background-origin: border-box;
+    transition: all .4s ease-in-out;
+    padding-top: 15px;
+    padding-bottom: 15px;
+}
+.demo-button:hover {
+    color: #7A76FF;
+}
