+++ apps/dash-lyft-explorer/assets/github/demo.gif
