+++ apps/dash-lyft-explorer/requirements.txt
@@ -1,13 +1,13 @@
-dash
-dash-deck
-dash-bootstrap-components
-lyft-dataset-sdk
+dash==2.4.1
+dash-deck==0.0.1
+dash-bootstrap-components==1.1.0
+pandas==1.4.2
+numpy==1.22.3
+colorlover==0.3.0
+lyft-dataset-sdk==0.0.8
+pydeck==0.7.1
+gunicorn==20.1.0
+
 pydeck==0.5.*
-jupyter
-colorlover
-plotly
 Pillow
-gunicorn
-numpy
-pandas
 opencv-python==4.2.0.34
