+++ apps/dash-lyft-explorer/runtime.txt
@@ -1,1 +1,1 @@
-python-3.7.6
+python-3.8.0
